Save labels as int32 (halves .npy size) and add run_bev.sh pipeline script
- main.py: cast labels to int32 before saving (0-7 class indices don't need
  int64, reduces labels.npy from ~18 GB to ~9 GB for 2.4B points)
- run_bev.sh: end-to-end script for BEV inference + apply labels + building
  counting, with configurable GPU, CPU affinity, and ramdisk output

+++ main.py
@@ -457,7 +457,7 @@ def process_pointcloud(
 
     # save labels
     _ensure_parent_dir(output_labels)
-    np.save(output_labels, all_labels)
+    np.save(output_labels, all_labels.astype(np.int32))
     print(f"Saved labels → {output_labels}")
 
     # optional colored PLY
